backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
google-cloud-vision==3.4.4
google-cloud-storage==2.10.0
google-cloud-bigquery==3.13.0
google-cloud-aiplatform==1.71.1
vertexai==1.71.1
pydantic==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
Pillow==10.1.0
pandas==2.1.4
numpy==1.24.3
requests==2.31.0
aiofiles==23.2.1
PyPDF2==3.0.1
